Updating tests.
The following 3 compilers should pass all of the tests:

Intel 17.0.4
GFortran 7.3
NAG 6.2

Many of the less exotic containers will work with other compiler
versions, but at this time these are the only ones known to work with
all combinations given in the test suite.   As of 7.3 GFortran now
correctly supports unlimited polymorphic pointers to deferred length strings.

There is an important caveat with regard to containers of POINTERs.
Set and altSet containers of pointers are in general suspect.  There
is no portable way within Fortran to _order_ pointers which is
essential for logarithmic search properties of Set containers.
However, a kludge with TRANSFER appears to be robustly portable
despite relying on non-guaranteed behavior

Further note: The tests for Set/altSet of pointers to unlimited
polymorphic entities are deactivated for GFortran.  (Once pFUnit
supports "ignored" tests, this will be handled in a more transparent
manner.)  Since the functionality is not guaranteed by the Fortran
standard, no bug report has been submitted.