Fix for bug with empty containers.
The value() method previously made illegal references when an iterator
is derived from an empty container.  (E.g., c%find(x)).

This was missed in the unit tests where the iterator was compared
against end().  But we expect value() to behave reasonably even in
this empty case.   Fixes for map, set, and alt_set containers.