Fixes #13.
This commit deactivates a bit of gFTL that breaks with ifort 18.  The
capability is for copying vectors and is mostly there to keep the interfaces
as close to STL as possible.

By default this commit does not change anything.  But if the user
specifies __INTEL_18 as an FPP token, then the offending logic is skipped
.