Fixes #25 - robust targets in containers.
When containers hold values with the ALLOCATABLE attribute, they
should be able support external pointers that remain valid as the
container changes.  The key is ensure that MOVE_ALLOC() is used in all
relevant cases instead of assignment.

For Vector containers this is mostly in place, but testing did reveal
a bug in the insert() method that is fixed here.


For altSet the functionality was also already present, apparently.
Presumably for Set it is trivial as Set uses pointers internally anway.

For (alt) Map, the functionality was latent.  Users must specify
_pair_allocatable so that altSet switches to MOVE_ALLOC()


Unfortunately, the truly desirable capability would be the case where
a map uses values that are ALLOCATABLE without regard to the key.
altSet would then need to MOVE_ALLOC() the value subcomponent of its
type, which is more intrusive than would otherwise be desirable.

And one might want key's to be preserved as well, at which point
_pair_allocatable is a fine solution.  So we will hold here for now.