Fixes #25 and Fixes #24.
Map now uses a custom __TYPE_MOVE for pair objects so that ALLOCATABLE
values are manipulated internally with move_alloc().

This also exposed the suspected but unproven theory that TARGET was
needed in a few more interfaces for pointers to be properly robust
under container changes.  (This requires the appropriate targets of
the pointer to be ALLOCATABLE and so cannot work on simple
containers.   Documentation will be "fun".)

Also modified flags to gFortran to handle longer lines that now appear
in some test cases.  Support for fixed format should be considered
deprecated.