Improved support for nested CMake projects.
The big problem (not affecting this repo) is a belated discovery that
having hyphens in project names interferes with find_package().
Convention is to use an env variable <project>_DIR, but env variables
cannot have hyphens.

But while I'm at it, I'm enforcing some conventions on capitalization.
CMake project names are all caps and library names are all lower case.