Various changes to test with pFUnit 4.0
Most changes are trivial, but the use of built-in cmake macros for
preprocessing required a bit of work.

Hardest aspect (conceptually) was to circumvent circular dependencies
among projects.  For this, a separate Cmake project "GFTL_TEST" was
introduced.  This allows find_package(pFUnit ...) to find the local
gFTL rather than the one it was built with.