Tightening and cleaning.
GFortran is complaining about the unlimited polymorphic (allocatable)
case, with run-time faults.  Added TARGET and constructor to all tests
in the hopes that it might help.

The problem appears to be a bit of a heisenbug though, as it does move
around a bit with print statements.  Current status still broken on OS
X with gfortran 8.2.  (I suspect earlier versions were broken in this
manner, but have not yet checked.)