Fixes #82 - Intel 18 memory leak workaround.
Intel 18 produces correct results, but does has a memory leak
associated with assignment of variables of derived type with deferred
length string components.

Making the workaround the default, but may reverse that if other
problems are detected downstream.   I have provided a new CPP token

__PAIR_ASSIGN(dest,src)

That can override this behavior.