Fixes #86 - double allocate in workaround
Previous release had a workaround for a memory leak in GFortran.
However for certain narrow use cases this workaround resulted in
code that attempted to reallocate an already allocated entity.

The situation is when a vector (and by implication altSet and altMap)
has an allocatable type, the set() method would allocate without
deallocating.  Use cases that simply grow and use containers without
updating existing entries were not affected.