Various workarounds for gFortran
Issues are mostly related to known bugs in using fhamcrest with
strings.  But there were similar issues is the unlimited polymorphic
case that are unrelated to teh use of fhamcrest.

The former were solved by resorting to older style assertions.  The
latter were resolved by not using strings in the test values for
containers of unlimited polymorphic entities.