Progress on introducing STL-like algorithms.
Just find() is integrated for the moment.  This exposed a problem in
defining EQ for deferred shape arrays.  Equality can be derived from
equality of elements, but one must first check the shapes of the 2
arrays.  And Fortran short-circuit rules (or lack thereof) prevents
this from just being a complicated inline expression.  Instead, a
local function needs to be defined with nested conditionals.  That is
deferred for now, and find() is not provided for this case.