Added in a few more test cases.
Cases involving Foo revealed a few minor bugs in the templating, but
also revealed serious problems with GFortran and related to fHamcrest
and the type `Foo` which was designed to work with fHamcrest.

For now have dumbed down Foo when using GFortran and this is requiring
lots of tests to be carefully edited to use the ASSERT macro instead
of `assert_that` when working with Foo.

All tests pass with NAG.