Make rc variable of map%at(key,rc) optional.
The map and ordered_map %at(key) function is mapped to the %of(key) function, whereas the %at(key,rc) is mapped to the %at_rc(key,rc) function

In other words:

val => map%at(non_existent_key, rc) uses the map%at_rc(key, rc) function.
and
val => map%at(non_existent_key) uses the map%of(key) function.

This has the suprising effect of searching for a non-existent key with map%at(key) to return a null-pointer as expected, BUT **also to add the non_existent_key to the map**, because it used the map%of(key) function under the hood.

This addition of the non-existent key is unexpected, and possibly a bug.

This situation has now been changed in this commit to make the %at(key) and %at(key, rc) functions both use the same implementation where the rc variable has been made optional.

Tests have been extended to cover finding a non-existent key with the %at(key) function.