skip move_alloc in vector and deque erase_range with empty iterator
Current code was calling MOVE_ALLOC(X, X) on all elements after the empty
iterator. Most compilers do not emit runtime error, but this is not legal
Fortran (see below), and the effect of most compiler is to deallocate X, which
is likely not the intention of erase_range(iterator, iterator) that I would
expect to be a no-op, while most compiler will compile it into code that
silently destroy all the vector elements after the iterator (but keep the
vector size unchanged).

Update the code to do nothing when the delta is not strictly positive (in both
case I do not think the code is meant to deal with reversed iterators, please
correct me if you think only the delta == 0 case should bypass the code).

Fortran 2018 section 16.9.137 point 4 mandates regarding MOVE_ALLOC(TO, FROM):
"if FROM and TO are the same variable, it shall be unallocated when MOVE_ALLOC
is invoked"